admin_app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0


pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

bcrypt==4.0.1
